# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Desk API",
    "version": "2.0.0",
    "description": "Desk business-management API — self-service email/password auth, business-setup drafts/businesses/memberships, and an admin table browser aggregating this service plus registry-api and compliance-os. Rewritten from the original Hono/Cloudflare Workers/D1 implementation onto Fastify/TypeScript/Postgres."
  },
...
  "components": {
    "securitySchemes": {
      "SessionToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque session token from POST /auth/signin, sent as Authorization: Bearer <token>."
      }
    },
...
  },
  "paths": {
...
    "/admin/tables": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List browsable tables (this service + registry-api + compliance-os)",
        "security": [
          {
            "SessionToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
...
  }
}